+++ degrees.cpp
@@ -5,19 +5,19 @@ namespace SSJ {
         this->degrees = 0.0;
     }
 
-    Degrees::Degrees(double degrees){
+    Degrees::Degrees(float degrees){
         this->degrees = fabs(fmod(degrees, 360.0));
     }
 
 
 
-    void Degrees::operator= (const double & degrees){
+    void Degrees::operator= (const float & degrees){
         this->degrees = fabs(fmod(degrees, 360.0));
     }
 
 
 
-    void Degrees::operator-= (const double & degrees){
+    void Degrees::operator-= (const float & degrees){
         this->degrees -= degrees;
         this->degrees = fabs(fmod(degrees, 360.0));
     }
@@ -29,7 +29,7 @@ namespace SSJ {
 
 
 
-    void Degrees::operator+= (const double & degrees){
+    void Degrees::operator+= (const float & degrees){
         this->degrees += degrees;
         this->degrees = fabs(fmod(degrees, 360.0));
     }
@@ -41,8 +41,8 @@ namespace SSJ {
 
 
 
-    bool Degrees::operator== (const double & degrees){
-       double temp =   fabs(fmod(degrees, 360.0));
+    bool Degrees::operator== (const float & degrees){
+       float temp =   fabs(fmod(degrees, 360.0));
         if(this->degrees == temp)
             return true;
         else
@@ -51,8 +51,8 @@ namespace SSJ {
 
 
 
-    bool Degrees::operator< (const double & degrees){
-        double temp =   fabs(fmod(degrees, 360.0));
+    bool Degrees::operator< (const float & degrees){
+        float temp =   fabs(fmod(degrees, 360.0));
         if(this->degrees < temp)
             return true;
         else
@@ -60,8 +60,8 @@ namespace SSJ {
     }
 
 
-    bool Degrees::operator> (const double & degrees){
-        double temp =   fabs(fmod(degrees, 360.0));
+    bool Degrees::operator> (const float & degrees){
+        float temp =   fabs(fmod(degrees, 360.0));
         if(this->degrees > temp)
             return true;
         else
@@ -69,8 +69,8 @@ namespace SSJ {
     }
 
 
-    bool Degrees::operator>= (const double & degrees){
-       double temp =   fabs(fmod(degrees, 360.0));
+    bool Degrees::operator>= (const float & degrees){
+       float temp =   fabs(fmod(degrees, 360.0));
         if(this->degrees >= temp)
             return true;
         else
@@ -79,15 +79,15 @@ namespace SSJ {
 
 
 
-    bool Degrees::operator<= (const double & degrees){
-        double temp =   fabs(fmod(degrees, 360.0));
+    bool Degrees::operator<= (const float & degrees){
+        float temp =   fabs(fmod(degrees, 360.0));
         if(this->degrees <= temp)
             return true;
         else
             return false;
     }
 
-    Degrees Degrees::operator+ (const double & degrees){
+    Degrees Degrees::operator+ (const float & degrees){
 		Degrees temp = *this;
         temp += degrees;
         temp.degrees = fabs(fmod(temp.degrees, 360.0));
@@ -95,7 +95,7 @@ namespace SSJ {
     }
 
 
-    Degrees & Degrees::operator- (const double & degrees){
+    Degrees & Degrees::operator- (const float & degrees){
         this-> degrees -= degrees;
         this->degrees = fabs(fmod(degrees, 360.0));
         return *this;
@@ -107,7 +107,7 @@ namespace SSJ {
         return *this ;
     }
 
-    double Degrees::getDegrees(){
+    float Degrees::getDegrees(){
         return this->degrees;
     }
 
@@ -119,7 +119,7 @@ namespace SSJ {
 
 
 
-    double Degrees::getRadians(){
+    float Degrees::getRadians(){
         return this->degrees * PI/180.0;
     }
