+++ dynamicobject.cpp
@@ -28,7 +28,7 @@ namespace SSJ{
     }
 
     void DynamicObject::RotateLeft(){
-        double deltaAngle = Config::RotationSpeed * DataContainer::DeltaTime.asSeconds();
+        float deltaAngle = Config::RotationSpeed * DataContainer::DeltaTime.asSeconds();
         bool a;
         if(this->targetAngle.getDegrees() - this->angle.getDegrees() < 0)
             a = true;
@@ -47,7 +47,7 @@ namespace SSJ{
     }
 
     void DynamicObject::RotateRight(){
-        double deltaAngle = Config::RotationSpeed * DataContainer::DeltaTime.asSeconds();
+        float deltaAngle = Config::RotationSpeed * DataContainer::DeltaTime.asSeconds();
         bool a;
         if(this->targetAngle.getDegrees() - this->angle.getDegrees() < 0)
             a = true;
@@ -70,9 +70,9 @@ namespace SSJ{
         Point newPosition = this->getMapPosition();
         SSJ::Degrees tempDegrees = this->angle;
         tempDegrees += addAngle;
-        double s = this->velocity * DataContainer::DeltaTime.asSeconds();
-        double px = sin(tempDegrees.getRadians()) * s;
-        double py = cos(tempDegrees.getRadians()) * s;
+        float s = this->velocity * DataContainer::DeltaTime.asSeconds();
+        float px = sin(tempDegrees.getRadians()) * s;
+        float py = cos(tempDegrees.getRadians()) * s;
         newPosition.x += px;
         newPosition.y -= py;
         return newPosition;
